+++ src/onelogin/saml2/settings.py
@@ -91,6 +91,7 @@ class OneLogin_Saml2_Settings(object):
         self.__contacts = {}
         self.__organization = {}
         self.__errors = []
+        self.__validate_sign = None
 
         self.__load_paths(base_path=custom_base_path)
         self.__update_paths(settings)
@@ -219,6 +220,7 @@ class OneLogin_Saml2_Settings(object):
             self.__security = settings.get('security', {})
             self.__contacts = settings.get('contactPerson', {})
             self.__organization = settings.get('organization', {})
+            self.__validate_sign = settings.get('validate_sign', None)
 
             self.__add_default_values()
             return True
@@ -332,6 +334,10 @@ class OneLogin_Saml2_Settings(object):
                 errors += self.check_idp_settings(settings)
             sp_errors = self.check_sp_settings(settings)
             errors += sp_errors
+
+        if 'validate_sign' in settings \
+                and not callable(settings['validate_sign']):
+            errors.append('validate_sign_not_a_function')
 
         return errors
 
@@ -795,3 +801,12 @@ class OneLogin_Saml2_Settings(object):
         :rtype: boolean
         """
         return self.__debug
+
+    def get_validate_sign(self):
+        """
+        Returns custom validate_sign method if specify or the default method
+
+        :return: a validate_sign method
+        :rtype: function
+        """
+        return self.__validate_sign or OneLogin_Saml2_Utils.validate_sign